backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
langgraph>=0.2.34
langchain>=0.3.7
langchain-community>=0.3.5
langchain-google-genai>=2.0.5
google-generativeai>=0.8.3
python-dotenv>=1.0.0
asyncpg>=0.29.0
pydantic>=2.9.2
httpx>=0.27.2
beautifulsoup4>=4.12.3
numpy>=1.26.4
redis>=5.2.0
psycopg2-binary>=2.9.9
aiofiles>=24.1.0
python-multipart>=0.0.6
lxml>=5.3.0